Support kubernetes login for vault
This would be a breaking change, as the configuration has to be updated.
However, the change is bringing retro-compatibility, so the old setup could
still be used for some time before we actually remove the support for it.

The internal config is also changed, but this shouldn't break the usage to
anyone, unless this type is used for some reason. All the major changes are done there and
inside the default VaultApi implementation

Signed-off-by: ivangonzalezacuna <ivangonzalezacuna@gmail.com>