Make adr search collator a backend-module
As it's already done to other collators, such as the techdocs, catalog or explore.
Let's make the ADR be usable for the new backend system with the new format.

I've used other available collators as inspiration, and added some module extensions
for those possible customizations that people might need if using different formats than MADR.

Signed-off-by: ivangonzalezacuna <ivangonzalezacuna@gmail.com>