Integrate notifications into the tech-insights backend
Make the notifications available as part of the FactRetriever context.
Now that this plugin is available, it would make sense to let users be
able to use this service to trigger notifications when a check is failing.

The usage of this service will be up to the end user, and no logic has to
be implemented into the tech-insights directly.

Signed-off-by: ivangonzalezacuna <ivangonzalezacuna@gmail.com>