ci: create auto-version-bump workflow
Adds a new GitHub Actions workflow (`auto-version-bump.yml`) to trigger
version bumps based on workspaces with a `.auto-version-bump` file.

`generate-auto-bump-matrix.js` generates a matrix of workspaces that contain
`.auto-version-bump` file, in batches of 10. Batches are triggered with a
5 minute delay to effort to avoid rate limit issues we hit before.

Initially, only the 'jfrog-artifactory' workspace has the `.auto-version-bump`
file. This is to test the workflow and ensure it works as expected before
adding to other workspaces.

Refs: https://github.com/backstage/community-plugins/issues/3231

Signed-off-by: Beth Griggs <bethanyngriggs@gmail.com>