fix(search-backend-module-confluence-collator): refine CQL query generation for spaces and query
Previously, the Confluence collator would generate an invalid CQL query
of the form `() and (label="test")` if the `spaces` configuration was
empty or not provided, but a `query` was specified. This resulted in
no search results.

This commit refines the CQL query generation logic to correctly handle
different combinations of `spaces` and `query` parameters:
- If only `spaces` are provided, the query will be `space=S1 or space=S2`.
- If only `query` is provided, the query will be `your_query_here`.
- If both `spaces` and `query` are provided, they will be combined
  as `(space="S1" or space="S2") and (your_query_here)`.
- If neither is provided, the default query to fetch all content types
  (`type IN (page, blogpost, comment, attachment)`) is used.

Additionally, new test cases have been added to cover these scenarios,
and the plugin documentation has been updated to clarify the behavior
of these parameters.

Signed-off-by: GitLab CI <128792991+dfl-aeb@users.noreply.github.com>