fix(rollbar-backend): refresh projects to evict expired tokens
Rollbar project metadata state is only flushed when the backend
service restarts. Rollbar project tokens can expire or be
deleted at the source of record resulting in invalid token
state in the plugin backend.

This cache is added as an optional dependency since the
createRouter function is still directly exposed to developers
as a public interface.

Signed-off-by: James Andrew Vaughn <jamesvaughn@modethirteen.com>