feat(argocd): Use K8s API, remove external dep (#5680)
* feat(argocd): Use K8s API, remove external dep

This change will standardize the plugin to use the K8s plugin API as
opposed to requiring an external dependency with a custom API name.

This is a breaking change that will require users of this plugin to now
set up the K8s plugin (frontend and backend) as a prerequisite.

This is similar to how the Tekton and Topology plugins work.

Signed-off-by: Bryan Ramos <bramos@redhat.com>

* Add RHDH config to readme

Signed-off-by: Bryan Ramos <bramos@redhat.com>

* Remove product-specific documentation

Signed-off-by: Bryan Ramos <bramos@redhat.com>

---------

Signed-off-by: Bryan Ramos <bramos@redhat.com>